Re-removed libtool macro
I mistakenly thought I disabled libtool in the first place.
It was actually just commented out, which frankly, I can't see why it'd be needed with cmake, so I've gone to the other extreme and deleted the macro that's not being used. (It's still in the commit history, but leaving around unused code is messy, and confusing as hopefully this PR demonstrates).