Fix paths in pkgconfig file
Commit 9e2e7178 tried to address an issue with wrong path assumptions
about relative vs. absolute directories and referenced a good explanation
how to do things correctly.

However, the commit itself does not follow the recommendations and this
leads to hard-coded full paths in the pkgconfig file - which is wrong
in my eyes and also breaks the build of dependend packages on e.g.
OpenWrt - I'd also call it regression since previous libzip releases
worked well.

This commit tries to fixup things by following the recommendation
of the already referenced page:
https://github.com/jtojnar/cmake-snips#concatenating-paths-when-building-pkg-config-files

Signed-off-by: Michael Heimpold <mhei@heimpold.de>