zip_mkstempm: Add flag to allow using a custom suffix
There could be such a scenario for new users:

1. Compress many files or large files.
2. A temporary file is created.
3. The user doesn't know it's a temporary file, and continuously tries to open it.

The situation has happened in a video: https://youtu.be/TtsglXhbxno?t=720

We should give the temporary file a descriptive suffix, and ".part" seems to be the
most reasonable and familiar suffix for most users, because the suffix
is also used by many browsers to indicate a file is still being downloaded.

To make the solution clean, _zip_mkstempm is also adapted to three use cases:

1. Default mode      (mkstemp)
2. Fill in name only (mktemp)
3. Fill in name only, but with a custom suffix

When calling _zip_stdio_op_create_temp_output, the third flag is used so
a custom suffix is made possible.